refactor(ollama): migrate auth proxy to .mts
Convert scripts/ollama-auth-proxy.js to a typed ESM .mts entrypoint running
under Node native type stripping without tsx. Preserve the fail-closed
Bearer-token check, byte-length gate before timingSafeEqual, authorization
header stripping, loopback backend, and EADDRINUSE exit. Trim the process
needle to ollama-auth-proxy so upgrade and recovery still detect the old .js
process next to the new .mts.
